requirements.txt
flask
faiss-cpu
PyMuPDF
python-docx
sentence-transformers
transformers
accelerate
torch
